refactor: Decompose monolithic SettingsScreen
- Extract `SettingsScreen` components into `ui/settings/components/`:
  - `SettingsGrid` and `ToggleSettingCard` moved to `SettingsGrid.kt`
  - `StepperSettingItem` moved to `StepperSettingItem.kt`
  - `SwitchSettingItem` moved to `SwitchSettingItem.kt`
  - Reduces `SettingsScreen.kt` from 309 to 111 lines (64% reduction).

- Move `SettingsScreenPreview` to the `debug` source set.
  - Created `app/src/debug/java/.../preview/SettingsScreenPreviews.kt`.
  - Improves separation between production and preview code.
